Refactor obtained date tooltip
